refactor: extract `createValidSession` helper function to setup file for improved session management
